Add DTOs for post requests

+++ src/components/scripts/mainHighscores.ts
@@ -1,69 +1,56 @@
-import {Component, Prop, Vue, Watch} from 'vue-property-decorator';
+import {Component, Prop, Vue} from 'vue-property-decorator';
 import axios from "axios";
 import Highscore from "../../components/models/Highscore";
-import UserData from "@/components/models/UserData";
+import UserDataDTO from "@/components/models/UserDataDTO";
+import HighscoreDTO from "@/components/models/HighscoreDTO";
 
 
 @Component
 export default class Highscores extends Vue {
     @Prop() private msg!: string;
     @Prop() private highscores!: Array<Highscore>;
-    @Prop() private apiUrl = 'http://localhost:3002';
-    @Prop() private userData!: UserData;
+    @Prop() private apiUrl = process.env.VUE_APP_HIGHSCORES_PATH;
+    @Prop() private userDataDTO!: UserDataDTO;
     @Prop() private time!: number;
     @Prop() private highScoresLimit = 10;
 
-    @Watch('time')
-    onPropertyChanged(){
-        console.log("wwwww");
-    }
-
     async created() {
-        console.log(process.env);
- /*       console.log("Highscores component created");
-        console.log(this.time);
-        console.log(this.userData.name);
-        console.log(this.userData.pictureUrl);
-        console.log(this.userData + "DUPSKo");
-*/
-        const hs = new Highscore(this.time, this.userData);
-        this.highscores=[];
-        this.updateHighscores();
-        this.sendNewHighscore(hs);
-        //console.log(this.time);
-        //this.updateHighscores();
-
+        //const hs = new HighscoreDTO(this.time, this.userDataDTO);
+        //this.highscores=[];
+        //await this.sendNewHighscore(hs);
+        //await this.updateHighscores();
     }
+
     async updateHighscores(){
         try {
-            const res = await axios.get(this.apiUrl);
+            const res = await axios.get(this.apiUrl+"/highscores");
             this.highscores = res.data;
             this.sortHighscores();
+
 
         } catch (e) {
             console.log("Error while receiving json data");
         }
     }
 
-    async sendNewHighscore(highscore: Highscore){
-        console.log("1");
-        if(this.isEligibleForHighscore(highscore)) {
-            console.log("2");
+    async sendNewHighscore(highscore: HighscoreDTO) {
+        if (highscore.Score > 0) {
+
+        if (this.isEligibleForHighscore(highscore)) {
             if (this.isHighscoresFull()) {
-                console.log("3");
                 await this.removeHighscoreById(this.getWorstHighScoreId());
             }
             await this.addHighscore(highscore);
         }
+    }
 
     }
-    isEligibleForHighscore(newHighscore: Highscore) {
+    isEligibleForHighscore(newHighscore: HighscoreDTO) {
         let hit = false;
-        console.log(this.highscores.length);
-        if(this.highscores.length===0){return true}
+        if(this.highscores.length<11){ return true }
 
         this.highscores.forEach((highscore: Highscore)=>{
-            if(newHighscore.score < highscore.score){ hit = true;}
+            if(newHighscore.Score < highscore.score){ hit = true }
         });
 
         return hit;
@@ -81,17 +68,19 @@ export default class Highscores extends Vue {
         });
         return worstHighscore.id;
     }
-     async addHighscore(highscore: Highscore){
-        console.log("4");
-        await axios.post(this.apiUrl+"/highscores", JSON.stringify(highscore));
+     async addHighscore(highscore: HighscoreDTO){
+        console.log("Adding highscore");
+        await axios.post(this.apiUrl+"/highscores", highscore);
     }
-    async removeHighscoreById(id: number){
-        await axios.delete(this.apiUrl + "/highscores/" + (id));
+    async removeHighscoreById(id: number | undefined){
+        if (!typeof(id)===undefined) {
+            await axios.delete(this.apiUrl + "/highscores/" + (id));
+        }
 
     }
     sortHighscores(){
         this.highscores.sort(function (a, b) {
-            return b.score - a.score;
+            return a.score - b.score;
         });
     }
     isHighscoresFull(){
